Don't create proxy classes every time they're instantiated by the DIC
This is subject to race conditions when multiple concurrent requests
are sent to a symfony project. This is because every request
regenerate the proxy class, thus by the time they're written and required
there might be another request rewriting the file, leading to class not found
or even EOF error in the middle of the proxy class.